+++ src/r0/account.rs
@@ -2,6 +2,7 @@
 
 pub mod change_password;
 pub mod deactivate;
+pub mod get_username_availability;
 pub mod register;
 pub mod request_password_change_token;
 pub mod request_register_token;